requirements.txt
openai
python-dotenv
requests
fastapi
uvicorn
python-multipart
pydantic
beautifulsoup4
PyPDF2
pandas
openpyxl
